SW-7216 - cleaning cache in BE deletes gitkeep

+++ engine/Shopware/Controllers/Backend/Cache.php
@@ -172,6 +172,9 @@ class Shopware_Controllers_Backend_Cache extends Shopware_Controllers_Backend_Ex
             );
 
             foreach ($iterator as $path) {
+                if(strcmp($path->getPathName(), (rtrim($cacheOptions['cache_dir'], '/').'/.gitkeep')) === 0) {
+                    continue;
+                }
                 if ($path->isDir()) {
                     rmdir($path->__toString());
                 } else {
